requirements.txt
click
pikepdf
ghostscript
Pillow